# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/lyrics/stream": {
      "get": {
        "description": "Server-Sent Events endpoint streaming real-time extraction, platform matching, and timing progress (60 RPM for cached hits, 6 RPM for live uncached resolutions).",
        "operationId": "TracksController_streamLyrics[0]",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Platform identifier (spotify, apple, deezer, netease, qq, isrc)",
            "schema": {
              "example": "spotify",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Platform-specific track ID",
            "schema": {
              "example": "4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "url",
            "required": false,
            "in": "query",
            "description": "Direct song/track URL from any streaming platform",
            "schema": {
              "example": "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Output lyric format (default: json)",
            "schema": {
              "enum": [
                "ttml",
                "lrc",
                "lrca2",
                "yrc",
                "qrc",
                "eslrc",
                "ass",
                "lyl",
                "lys",
                "lqe",
                "json"
              ],
              "type": "string"
            }
          },
          {
            "name": "forceRefresh",
            "required": false,
            "in": "query",
            "description": "Bypass cached database results and force live re-fetching",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream emitting lifecycle events: init, cache_hit, cache_miss, resolving, platform_matched, lyrics_searching, lyrics_found, saving, done, error",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stage": {
                      "type": "string",
                      "example": "done"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1787262680794
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameters or invalid format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Stream lyrics resolution progress via SSE",
        "tags": [
          "Tracks",
          "Lyrics"
        ]
      }
    },
    "/lyrics/stream": {
      "get": {
        "description": "Server-Sent Events endpoint streaming real-time extraction, platform matching, and timing progress (60 RPM for cached hits, 6 RPM for live uncached resolutions).",
        "operationId": "TracksController_streamLyrics[1]",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Platform identifier (spotify, apple, deezer, netease, qq, isrc)",
            "schema": {
              "example": "spotify",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Platform-specific track ID",
            "schema": {
              "example": "4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "url",
            "required": false,
            "in": "query",
            "description": "Direct song/track URL from any streaming platform",
            "schema": {
              "example": "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Output lyric format (default: json)",
            "schema": {
              "enum": [
                "ttml",
                "lrc",
                "lrca2",
                "yrc",
                "qrc",
                "eslrc",
                "ass",
                "lyl",
                "lys",
                "lqe",
                "json"
              ],
              "type": "string"
            }
          },
          {
            "name": "forceRefresh",
            "required": false,
            "in": "query",
            "description": "Bypass cached database results and force live re-fetching",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream emitting lifecycle events: init, cache_hit, cache_miss, resolving, platform_matched, lyrics_searching, lyrics_found, saving, done, error",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stage": {
                      "type": "string",
                      "example": "done"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1787262680794
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameters or invalid format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Stream lyrics resolution progress via SSE",
        "tags": [
          "Tracks",
          "Lyrics"
        ]
      }
    },
    "/api/tracks/stream": {
      "get": {
        "description": "Alias for /api/lyrics/stream",
        "operationId": "TracksController_streamTracks[0]",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Platform identifier (spotify, apple, deezer, netease, qq, isrc)",
            "schema": {
              "example": "spotify",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Platform-specific track ID",
            "schema": {
              "example": "4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "url",
            "required": false,
            "in": "query",
            "description": "Direct song/track URL",
            "schema": {
              "example": "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Output lyric format (default: json)",
            "schema": {
              "enum": [
                "ttml",
                "lrc",
                "lrca2",
                "yrc",
                "qrc",
                "eslrc",
                "ass",
                "lyl",
                "lys",
                "lqe",
                "json"
              ],
              "type": "string"
            }
          },
          {
            "name": "forceRefresh",
            "required": false,
            "in": "query",
            "description": "Bypass cached results and force live re-fetching",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream emitting real-time progress events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stage": {
                      "type": "string",
                      "example": "done"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1787262680794
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Stream track resolution progress via SSE",
        "tags": [
          "Tracks"
        ]
      }
    },
    "/tracks/stream": {
      "get": {
        "description": "Alias for /api/lyrics/stream",
        "operationId": "TracksController_streamTracks[1]",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Platform identifier (spotify, apple, deezer, netease, qq, isrc)",
            "schema": {
              "example": "spotify",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Platform-specific track ID",
            "schema": {
              "example": "4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "url",
            "required": false,
            "in": "query",
            "description": "Direct song/track URL",
            "schema": {
              "example": "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Output lyric format (default: json)",
            "schema": {
              "enum": [
                "ttml",
                "lrc",
                "lrca2",
                "yrc",
                "qrc",
                "eslrc",
                "ass",
                "lyl",
                "lys",
                "lqe",
                "json"
              ],
              "type": "string"
            }
          },
          {
            "name": "forceRefresh",
            "required": false,
            "in": "query",
            "description": "Bypass cached results and force live re-fetching",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream emitting real-time progress events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stage": {
                      "type": "string",
                      "example": "done"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1787262680794
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Stream track resolution progress via SSE",
        "tags": [
          "Tracks"
        ]
      }
    },
    "/api/tracks": {
      "get": {
        "description": "Looks up a track in the database cache (~2ms, 60 RPM/IP) or fetches and resolves metadata upstream (6 RPM/IP).",
        "operationId": "TracksController_getTrack[0]",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Platform identifier (spotify, apple, deezer, netease, qq, isrc)",
            "schema": {
              "example": "spotify",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Platform-specific track ID",
            "schema": {
              "example": "4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          },
          {
            "name": "url",
            "required": false,
            "in": "query",
            "description": "Direct song/track URL from Spotify, Apple Music, Deezer, NetEase, or QQ Music",
            "schema": {
              "example": "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sanitized track metadata with cross-platform IDs and lyric availability flag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanitizedTrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing url or platform + id parameters or invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Track could not be resolved from upstream platforms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get or synchronize a track by platform, ID, or URL",
        "tags": [
          "Tracks"
        ]
      },
      "post": {
        "description": "Alternative JSON body endpoint to look up (60 RPM/IP) or synchronize (6 RPM/IP) track metadata.",
        "operationId": "TracksController_postTrack[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sanitized track metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanitizedTrackDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Lookup/sync a track via POST body",
        "tags": [
          "Tracks"
        ]
      }
    },
...
  },
  "info": {
    "title": "LyricsDB API",
    "description": "LyricsDB is a high-performance open-source lyrics indexing and synchronization platform.\n\n### Features\n* **Multi-Platform Resolution**: Cross-references Spotify, Apple Music, Deezer, NetEase Cloud Music, QQ Music, and ISRC.\n* **Precision Timed Lyrics**: Word-by-word and syllable-level sync with background vocals and duet roles.\n* **Format Engine**: Real-time conversion to TTML, LRC, LRCA2, YRC, QRC, ESLRC, ASS, LYL, LYS, LQE, and compact JSON.\n* **Real-Time Streaming**: Server-Sent Events (SSE) tracking extraction lifecycle and timing progress.\n* **PostgreSQL Fast Cache**: Sub-millisecond indexed read-through caching with thundering-herd deduplication.\n* **Provider Rate-Limiting & Circuit Breakers**: Smart queueing, RPM limits, and automated cooldowns to protect upstream services.",
    "version": "1.0.0",
    "contact": {
      "name": "LyricsDB Support",
      "url": "https://github.com/Anson10124/LyricsDB",
      "email": "support@lyricsdb.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current Host / Reverse Proxy (Auto-detected)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP response status code"
          },
          "message": {
            "type": "string",
            "example": "Please provide either \"url\" or both \"platform\" and \"id\".",
            "description": "Detailed explanation of the error"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "HTTP error title / type name"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
    }
  },
...
}